Make PlotlyJS/Plots weak dependencies to fix the JSON version conflict
PlotlyJS 0.18.x requires JSON.jl 0.20-0.21 (the pre-1.0 line); anything
needing modern JSON.jl (e.g. Oxygen.jl, which ExcelAtRetail.jl's backend
depends on for its HTTP API) needs 1.x. No single JSON version satisfies
both, so any environment depending on both this package and something
needing modern JSON was unresolvable outright, not just at plot time -
hit for real trying to `Pkg.instantiate()` ExcelAtRetail.jl's backend.

PlotlyJS/Plots were unconditional dependencies only because Visualization.jl
was unconditionally `include`d - nothing about the actual optimization/
querying functionality needs them. Split into two package extensions
instead, matching the granularity test/Project.toml already implied (it
only ever had PlotlyJS, never Plots):

- ext/SupplyChainOptimizationPlotlyJSExt.jl (triggered by PlotlyJS alone):
  plot_inventory, plot_network, plot_costs, plot_financials, plot_flows,
  animate_network, animate_flows - everything that doesn't touch Plots.jl.
- ext/SupplyChainOptimizationPlotsExt.jl (triggered by PlotlyJS + Plots):
  just movie_network, the one function needing Plots.Animation/mp4.

get_financials moved to Querying.jl - it never touched PlotlyJS/Plots (pure
DataFrame construction), so it stays available without either extension.
src/Visualization.jl now holds just docstring'd stub declarations
(`function name end`) so the names stay part of the package's exported API
surface with documentation regardless of whether the extensions are active.

Also fixed what looks like a pre-existing latent bug while relocating
movie_network: it referenced `Plots.Animation`/`Plots.mp4` (qualified) but
only ever `using Plots: Animation, buildanimation, mp4` (selective), which
doesn't bind the `Plots` name itself - likely never caught because
movie_network isn't covered by any test file, and can't be with plain
PlotlyJS in test/Project.toml. Now calls the already-unqualified names
directly. Added `using PlotlyJS` to test/runtests.jl so the PlotlyJS
extension actually activates during CI - without it every plot_*/animate_*
call in the test suite would be a fresh MethodError.

Also drops the [sources] override for SupplyChainModeling, now that it's
a normal registered dependency (General v0.4.0+) - see the previous
release's commit for that history.

Version bump: 0.4.1 -> 0.5.0. Under this package's pre-1.0 versioning,
this counts as breaking (a caller doing bare `using SupplyChainOptimization`
without also loading PlotlyJS/Plots would previously get working plot_*
functions and now gets a MethodError until they also load PlotlyJS/Plots) -
flagging it as such for AutoMerge's release notes requirement, even though
this is very unlikely to affect real callers (a plotting function is
useless without something to render the plot).

Note: not run against a real Julia install - no Julia runtime was
available in the environment this was written in. Opening as a PR rather
than merging directly (unlike the previous compat-only fixes) so this
repo's own CI actually exercises it before it goes to main/gets
registered.
