+++ Project.toml
@@ -1,18 +1,24 @@
 name = "SupplyChainOptimization"
 uuid = "785807e5-a5cf-4c87-8002-234621def379"
 authors = ["SupplyChef <47842426+SupplyChef@users.noreply.github.com>"]
-version = "0.4.1"
+version = "0.5.0"
 
 [deps]
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
-PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 SupplyChainModeling = "8a0b78ce-42db-494b-a45c-80b287de1a3d"
+
+[weakdeps]
+PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+
+[extensions]
+SupplyChainOptimizationPlotlyJSExt = "PlotlyJS"
+SupplyChainOptimizationPlotsExt = ["PlotlyJS", "Plots"]
 
 [compat]
 Dates = "1.10"
@@ -26,15 +32,7 @@ StatsBase = "0.34"
 SupplyChainModeling = "0.2, 0.3, 0.4"
 julia = "1.10"
 
-# TEMPORARY: the maximum_units/overflow_unit_cost support this depends on has
-# already merged to SupplyChainModeling's main (now at 0.4.0 - 0.4.0 only
-# added MaturationSource/QuotaSink, additive and unused here, so this is
-# just riding along), but isn't published to a registry yet, so point at
-# main directly rather than a moving feature branch. Remove this section
-# once a registry release exists - and keep the compat bound above in sync
-# with whatever version main currently declares in the meantime, or
-# [sources]-based resolution fails outright (main's declared version must
-# satisfy this bound; unlike a plain `Pkg.add(rev=...)`, [sources] does not
-# bypass compat checking).
-[sources]
-SupplyChainModeling = {url = "https://github.com/SupplyChef/SupplyChainModeling.jl", rev = "main"}
+# SupplyChainModeling is a normal registered dependency (General, v0.4.0+)
+# as of 2026-09-02 - no [sources] override needed. It briefly needed one
+# (the registry had drifted behind `main`) - see PLAN.md in ExcelAtRetail.jl
+# and this repo's git history around that date if that situation recurs.
